Removed dbTypes.py Now declare dbTypes within sourceDatabase.py
