--- src/lib/sourceObjs/dbTypes.py
@@ -1,8 +1,0 @@
-from enum import Enum, auto
-
-class DBType(Enum):
-    UNKNOWN = auto()
-    SPECIFIC = auto()
-    LOCATION = auto()
-    SCRIPTURL = auto()
-    SCRIPTDATA = auto()
